Fixing a bug with the soldier's dealDamage function (was fetching the wrong position) as well as with the use of the ceiling function (appropriately added a cast to double).

+++ pt2/part_b/Sniper.cpp
@@ -34,7 +34,7 @@ namespace mtm
     {
        units_t distance = GridPoint::distance(getPosition(), target);
 
-       if(distance < ceil(getRange() / DISTANCE_FACTOR) || distance > this->getRange())
+       if(distance < ceil((double)getRange() / DISTANCE_FACTOR) || distance > this->getRange())
        {
            throw OutOfRange();
        }
